test: move `println` tests into integration tests
So each test is compiled as separate binary and can get its own
`DebugLogger`
