+++ tests/integration/Cargo.toml
@@ -25,7 +25,7 @@ miden-mast-package.workspace = true
 miden-protocol = { workspace = true, features = ["std"] }
 miden-standards = { workspace = true, features = ["std"] }
 miden-processor.workspace = true
-miden-debug = { workspace = true, features = ["proptest"] }
+miden-debug = { workspace = true, features = ["proptest", "tui"] }
 midenc-dialect-arith.workspace = true
 midenc-dialect-cf.workspace = true
 midenc-dialect-hir.workspace = true
