+++ tests/integration/tests/debug_executor_println_dynamic.rs
@@ -1,34 +1,16 @@
+use log::Level;
 use miden_core::Felt;
+use miden_debug::logger::DebugLogger;
+use miden_integration_tests::{CompilerTest, testing::eval_package};
 use midenc_frontend_wasm::WasmTranslationConfig;
-
-use crate::{CompilerTest, testing::eval_package};
-
-#[test]
-fn test_println_static() {
-    let main_fn = "() -> u32 { println(\"hello\"); 0 }";
-    let mut test = CompilerTest::rust_fn_body_with_sdk(
-        "test_println",
-        main_fn,
-        WasmTranslationConfig::default(),
-        [],
-    );
-
-    let package = test.compile_package();
-
-    let args = [];
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let result: Felt = trace.parse_result().unwrap();
-        assert_eq!(result, Felt::from(0u32));
-        assert_eq!(trace.printed_lines().len(), 1);
-        assert_eq!(trace.printed_lines().values().next().unwrap(), "hello");
-        Ok(())
-    })
-    .unwrap();
-}
 
 // Manipulates bytes to avoid `call_indirect`, which gets triggered by Rust's formatting infra.
 #[test]
 fn test_println_dynamic() {
+    DebugLogger::init_for_tests()
+        .expect("each test using DebugLogger should run in its own process");
+    log::set_max_level(log::LevelFilter::Warn);
+
     let main_fn = r#"(digit: u32) -> u32 {
         let mut s = alloc::string::String::from("the digit is: ");
         // b'0' is zero as ASCII, then add `digit` as offset
@@ -44,14 +26,27 @@ fn test_println_dynamic() {
     );
 
     let package = test.compile_package();
+    let before = DebugLogger::get().clone_captured().len();
+    log::set_max_level(log::LevelFilter::Info);
 
     let args = [Felt::from(7u32)];
     eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
         let result: Felt = trace.parse_result().unwrap();
         assert_eq!(result, Felt::from(0u32));
-        assert_eq!(trace.printed_lines().len(), 1);
-        assert_eq!(trace.printed_lines().values().next().unwrap(), "the digit is: 7");
         Ok(())
     })
     .unwrap();
+
+    let logs: Vec<_> = DebugLogger::get().clone_captured().into_iter().skip(before).collect();
+    let info_messages: Vec<_> = logs
+        .iter()
+        .filter(|entry| entry.level == Level::Info)
+        .map(|entry| entry.message.as_str())
+        .collect();
+    assert_eq!(
+        info_messages.as_slice(),
+        ["the digit is: 7"],
+        "observed logs: {:?}",
+        logs.iter().map(|e| format!("{}: {}", e.level, e.message)).collect::<Vec<_>>(),
+    );
 }
